Added speaker test using external DAC
Added separate module test for speaker with MCP4728 DAC as driver.
Test is meant to drive a 440 Hz sinusoid on the speaker via the DAC.
Preliminary results are promising, as there is a tone *but* it's not the
best quality, likely due to how the sinusoid value is calculated.

Still, shows a proof-of-concept of using MCP4728 to drive speaker.

Also trims unused pytest.