Added prints for stack/heap usage, lowered stacks
Added printout for the stack usage of all tasks, as well as the
remaining heap, to the stats task printout.

Stack watermark seems accurate for the DSP, Rx/Tx tasks, but highly
inaccurate for the stats task based on experimentation.
Also lowered stack sizes for all tasks based on highest watermarks.