Added approx output inst freq
Used the same principal to estimate the input instantaneous frequency to
calculate an estimate of the output instantaneous frequency.
I.e., use the previous frame phase to calculate the phase shift
necessary to result in the desired actual frequency shift.

Results are somewhat inconclusive, and subjective. Pitch appears to be
held *somewhat* better, but it's slight at best.
Calculation time is now also much higher.